Add loading state to PokemonListScreen

+++ app/src/main/java/com/example/testpockemonapp/ui/screens/pokemon_list/PokemonListScreen.kt
@@ -1,17 +1,20 @@
 package com.example.testpockemonapp.ui.screens.pokemon_list
 
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.Favorite
 import androidx.compose.material.icons.filled.FavoriteBorder
 import androidx.compose.material3.Card
 import androidx.compose.material3.CenterAlignedTopAppBar
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
@@ -21,16 +24,18 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
+import androidx.paging.LoadState
 import androidx.paging.PagingData
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.example.testpockemonapp.domain.model.PokemonListItem
+import com.example.testpockemonapp.ui.composables.shimmerLoading
 import kotlinx.coroutines.flow.flowOf
-import java.util.UUID
 
 @Composable
 fun PokemonListScreen(
@@ -63,27 +68,52 @@ private fun PokemonListContent(
         },
         content = {
             Box(modifier = modifier.padding(it)) {
-                LazyColumn(modifier = Modifier.fillMaxSize()) {
-                    items(
-                        count = lazyPagingItems.itemCount,
-                        key = { index ->
-                            val item = lazyPagingItems[index]
-                            item?.name.orEmpty()
-                        },
-                        itemContent = { index ->
-                            val item = lazyPagingItems[index]
+                when (lazyPagingItems.loadState.refresh) {
+                    is LoadState.Loading -> {
+                        ListLoadingState()
+                    }
 
-                            if (item != null) {
-                                PokemonListItemCard(
-                                    modifier = Modifier.animateItem(),
-                                    pokemon = item,
-                                    onItemClick = { onItemClick(item.name) },
-                                    onFavoriteIconClick = { onFavoriteIconClick(item.name) },
-                                    onDeleteIconClick = { onDeleteIconClick(item.name) }
-                                )
+                    else -> {
+                        LazyColumn(modifier = Modifier.fillMaxSize()) {
+                            items(
+                                count = lazyPagingItems.itemCount,
+                                key = { index ->
+                                    val item = lazyPagingItems[index]
+                                    item?.name.orEmpty()
+                                },
+                                itemContent = { index ->
+                                    val item = lazyPagingItems[index]
+
+                                    if (item != null) {
+                                        PokemonListItemCard(
+                                            modifier = Modifier
+                                                .padding(
+                                                    vertical = 8.dp,
+                                                    horizontal = 16.dp
+                                                )
+                                                .animateItem(),
+                                            pokemon = item,
+                                            onItemClick = { onItemClick(item.name) },
+                                            onFavoriteIconClick = { onFavoriteIconClick(item.name) },
+                                            onDeleteIconClick = { onDeleteIconClick(item.name) }
+                                        )
+                                    }
+                                }
+                            )
+
+                            if (lazyPagingItems.loadState.append is LoadState.Loading) {
+                                item {
+                                    Box(modifier = Modifier.fillMaxWidth()) {
+                                        CircularProgressIndicator(
+                                            Modifier
+                                                .align(Alignment.Center)
+                                                .padding(8.dp)
+                                        )
+                                    }
+                                }
                             }
                         }
-                    )
+                    }
                 }
             }
         }
@@ -99,9 +129,7 @@ private fun PokemonListItemCard(
     modifier: Modifier = Modifier,
 ) {
     Card(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(vertical = 8.dp, horizontal = 16.dp),
+        modifier = modifier.fillMaxWidth(),
         onClick = onItemClick
     ) {
         Row(
@@ -128,6 +156,24 @@ private fun PokemonListItemCard(
             IconButton(onClick = onDeleteIconClick) {
                 Icon(imageVector = Icons.Default.Delete, contentDescription = null)
             }
+        }
+    }
+}
+
+@Composable
+fun ListLoadingState(modifier: Modifier = Modifier) {
+    Column(modifier = modifier.fillMaxSize()) {
+        repeat(10) {
+            PokemonListItemCard(
+                modifier = Modifier
+                    .padding(vertical = 8.dp, horizontal = 16.dp)
+                    .clip(RoundedCornerShape(16.dp))
+                    .shimmerLoading(),
+                pokemon = PokemonListItem("loading", "", false),
+                onItemClick = {},
+                onFavoriteIconClick = {},
+                onDeleteIconClick = {},
+            )
         }
     }
 }
